패들과 블록의 공통부분을 추상클래스로 작성
