+++ yongjae/src/main/java/com/nhnacademy/Paddle.java
@@ -1,30 +1,15 @@
 package com.nhnacademy;
 
-import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
-public class Paddle {
-    private double x; // 패들의 중심 x 좌표
-    private double y; // 패들의 중심 y 좌표
-    private double width; // 패들의 너비
-    private double height; // 패들의 높이
+public class Paddle extends Rectangle {
     private double speed; // 패들의 이동 속도
-    private Color color; // 패들의 색상
 
     // 생성자
     public Paddle(double x, double y, double width, double height, double speed, Color color) {
-        this.x = x;
-        this.y = y;
-        this.width = width;
-        this.height = height;
+        super(x, y, width, height);
         this.speed = speed;
         this.color = color;
-    }
-
-    // 패들을 그리는 메서드
-    public void draw(GraphicsContext gc) {
-        gc.setFill(color);
-        gc.fillRect(x - width / 2, y - height / 2, width, height); // 중심을 기준으로 사각형 그리기
     }
 
     // 패들의 위치를 왼쪽으로 이동
@@ -37,6 +22,7 @@ public class Paddle {
         x += speed;
     }
 
+    @Override
     public boolean checkCollision(Ball ball) {
         // 공이 패들의 경계와 충돌했는지 확인
         return ball.getX() + ball.getRadius() > x &&
@@ -47,31 +33,15 @@ public class Paddle {
 
     // 패들이 화면 경계를 벗어나지 않도록 제한
     public void checkBounds(double canvasWidth) {
-        if (x - width / 2 < 0) { // 왼쪽 경계
-            x = width / 2;
-        } else if (x + width / 2 > canvasWidth) { // 오른쪽 경계
-            x = canvasWidth - width / 2;
+        if (x < 0) { // 왼쪽 경계
+            x = 0;
+        } else if (x + width > canvasWidth) { // 오른쪽 경계
+            x = canvasWidth - width;
         }
     }
 
     // Getter와 Setter (필요 시 사용)
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-    public double getWidth() {
-        return width;
-    }
-
     public void setWidth(double width) {
         this.width = width;
-    }
-
-    public double getHeight() {
-        return height;
     }
 }
